pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Michał Grela <michal.grela@themomentum.ai>", "Filip Begiełło <filip.begiello@themomentum.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.115.6"
pydantic = "^2.7.0"
uvicorn = "^0.30.0"
llama-index = "^0.12.2"
llama-index-llms-ollama = "^0.4.2"
whisperx = "^3.3.1"
llvmlite = "^0.43.0"
numba = "^0.60.0"
ollama = "^0.4.4"
exceptiongroup = "^1.1.0"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.20"

[tool.poetry.group.demo.dependencies]
gradio = "^5.12.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-cov = "^5.0.0"
autoflake = "^2.3.1"
flake8 = "^7.0.0"
mypy = "^1.10.0"
isort = "^5.13.2"
pre-commit = "^3.7.1"
black = "^24.4.2"
types-requests = "^2.32.0.20241016"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
